cloud-functions/ai-services/requirements.txt
google-cloud-vision==3.4.4
google-cloud-language==2.11.1
google-cloud-translate==3.11.3
google-cloud-aiplatform==1.36.4
flask==2.3.3
flask-cors==4.0.0
functions-framework==3.4.0
